social_network/requirements.txt
python-dotenv==0.20.0
Django==4.0.4
django-bootstrap4==22.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-simple-captcha==0.5.17
Pillow==9.1.0
psycopg2-binary
celery==5.2.6
eventlet==0.33.1
channels==3.0.5
channels-redis==3.4.1
asgiref==3.5.0
django-user-agents==0.4.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.1
pyjwt==2.5.0
drf-yasg==1.21.4
pyTelegramBotAPI==4.7.0
djet==0.3.0
gunicorn==20.1.0